Admmin dashboard creation with approval for user post and account

+++ blog_app/views.py
@@ -1,32 +1,29 @@
-from django.shortcuts import render, redirect
-from django.contrib.auth.forms import UserCreationForm
-from django.contrib.auth import login, logout, authenticate
+from django.shortcuts import render, redirect, get_object_or_404
+from django.contrib.auth import get_user_model, login, logout, authenticate
 from django.contrib.auth.models import User
 from django.contrib import messages
-from django.contrib.auth import password_validation
+from django.contrib.auth.password_validation import validate_password
 from django.contrib.auth.decorators import login_required
+from django.core.exceptions import PermissionDenied
+from django.db import transaction
 
-from .models import Blog, UserProfileList, Comment
+from .models import Blog, Profile, Comment
 from .forms import *
 from .validators import *
+from .decorators import unautorized_access, admin_only
+
 # Create your views here.
 @login_required(login_url="log-in")
 def home(request):
-    try:
-        profile = UserProfileList.objects.all()
-        blog = Blog.objects.all().order_by('?')
-        comment = Comment.objects.all()
-        cont_dict = {
-            "profile" : profile,
-            "blog" : blog,
-            "comment" : comment,
-        }
-        return render(request, "blog_app/index.html", cont_dict)
-    except Blog.DoesNotExist:
-        messages.error(request, "Create a blog first!")
-        return render(request, "blog_app/index.html")
+    blogs = (Blog.objects
+             .select_related('author', 'author__profile')
+             .prefetch_related('comments__author__profile')
+             .order_by('-created_at')
+             .filter(approved=True))
+    return render(request, "blog_app/index.html", {"blog" : blogs})
 
 #------------------- Authentication ----------------------
+@unautorized_access
 def registration(request):
     form = CustomUserCreationForm(request.POST or None)
     if form.is_valid():
@@ -38,14 +35,14 @@ def registration(request):
             try:
                 check_username(username)
                 email_validate(email)
-                password_validation(password)
+                validate_password(password)
                 password_validate(password)
                 
-                User.objects.create_user(username=username,email=email,password=password)
-                messages.success(request, "Account successfully created")
-            
-                user = User.objects.get(username=username)
-                UserProfileList.objects.create(profile_id=user.id, email=user.email)
+                User = get_user_model()
+                with transaction.atomic():
+                    user = User.objects.create_user(username=username,email=email,password=password)
+                    Profile.objects.create(user=user, email=user.email)
+                    
                 auth = authenticate(request, username=username, password=password)
                 if auth is not None:
                     login(request, user)
@@ -58,6 +55,7 @@ def registration(request):
             return redirect("registration")
     return render(request, "authenticate/registration.html", {"form":form})
 
+@unautorized_access
 def log_in(request):
     form = UserLoginForm(request.POST or None)
     if form.is_valid():
@@ -66,18 +64,32 @@ def log_in(request):
         
         user = authenticate(request, username=username, password=password)
         if user is not None:
-            login(request, user)
-            messages.success(request, "Login Successfull!")
-            return redirect("home")
+            if not user.is_superuser:
+                profile = Profile.objects.get(user_id=user.id)
+                if profile.is_active:
+                    login(request, user)
+                    messages.success(request, "Login Successfull!")
+                    return redirect("home")
+                else:
+                    messages.error(request, "Your account has been blocked")
+                    return redirect("log-in")
+            else:
+                login(request, user)
+                messages.success(request, "Login Successfull!")
+                return redirect("home")
+        else:
+            messages.error(request, "Username or Password doesn't match")
+            return redirect("log-in")
     return render(request, "authenticate/log_in.html", {"form":form})
 
+@login_required(login_url="log-in")
 def log_out(request):
     logout(request)
     return redirect("log-in")
 
 @login_required(login_url="log-in")
 def create_profile(request):
-    user = UserProfileList.objects.get(profile_id=request.user.id)
+    user = Profile.objects.get(user_id=request.user.id)
     form = ProfileCreationForm(instance=user)
     if request.method == "POST":
         form = ProfileCreationForm(request.POST, request.FILES, instance=user)
@@ -103,35 +115,33 @@ def change_password(request):
 #---------------- Blog Section ----------------------
 @login_required(login_url="log-in")
 def create_blog(request):
-    form = BlogCreationForm()
-    if request.method == "POST":
-        form = BlogCreationForm(request.POST, request.FILES)
-        if form.is_valid():
-            form.save(commit=False)
-            title = form.cleaned_data['title']
-            content = form.cleaned_data['content']
-            image = form.cleaned_data['blog_image']
-            
-            Blog.objects.create(name_id=request.user.id, title=title, content=content, blog_image=image)
-            messages.success(request, "Blog Successfully created!")
-            return redirect("view-profile")
+    form = BlogCreationForm(request.POST or None, request.FILES or None)
+    if form.is_valid():
+        post = form.save(commit=False)
+        post.author = request.user
+        post.save()
+        messages.success(request, "Blog Successfully created!")
+        return redirect("view-profile")
     return render(request, "blog_app/create_blog.html", {"form" : form})
 
 @login_required(login_url="log-in")
 def update_blog(request, id):
-    blog = Blog.objects.get(id=id)
-    form = BlogCreationForm(instance=blog)
-    if request.method == "POST":
-        form = BlogCreationForm(request.POST, request.FILES, instance=blog)
-        if form.is_valid():
-            form.save()
-            messages.success(request, "Blog Successfully Updated!")
-            return redirect("view-profile")
-    return render(request, "blog_app/update_blog.html", {"form" : form})
+    blog = get_object_or_404(Blog, pk=id)
+    if blog.author != request.user and not request.user.is_staff:
+        raise PermissionDenied("Not allowed to edit")
+    form = BlogCreationForm(request.POST or None, request.FILES or None, instance=blog)
+    if form.is_valid():
+        form.save()
+        messages.success(request, "Blog updated")
+        return redirect("view-profile")
+    return  render(request, "blog_app/update_blog.html", {"form" : form})
 
 @login_required(login_url="log-in")
 def delete_blog(request, id):
-    blog = Blog.objects.get(id=id)
+    # blog = Blog.objects.get(id=id)
+    blog = get_object_or_404(Blog, pk=id)
+    if blog.author != request.user and not request.user.is_staff:
+        raise PermissionDenied("Not allowed to delete this blog")
     blog.delete()
     messages.success(request, "Blog successfully deleted!")
     return redirect("view-profile")
@@ -139,23 +149,66 @@ def delete_blog(request, id):
 @login_required(login_url="log-in")
 def view_profile(request):
     try:
-        prfoile = UserProfileList.objects.get(profile_id=request.user.id)
-        blog = Blog.objects.filter(name_id=request.user.id)
+        blog = Blog.objects.filter(author_id=request.user.id).select_related('author', 'author__profile').prefetch_related('comments__author__profile').order_by('-created_at')
         cont_dict = {
-            "profile" : prfoile,
             "blog" : blog
         }
         return render(request, "blog_app/view_profile.html", cont_dict)
-    except UserProfileList.DoesNotExist:
+    except Blog.DoesNotExist:
         messages.error(request, "Create your profile first")
         return redirect("home")
     
 @login_required(login_url="log-in")
 def blog_comment(request, id):
+    blog = get_object_or_404(Blog.objects.select_related('author'), pk=id)
     form = CommentForm(request.POST or None)
     if form.is_valid():
-        comment = form.cleaned_data['comments']
-        Comment.objects.create(blog_id=id, commenter_id = request.user.id, comments=comment)
-        messages.success(request, "Success")
+        with transaction.atomic():
+            Comment.objects.create(blog=blog, author=request.user, comments=form.cleaned_data["comments"])
+        messages.success(request, "Comment added")
         return redirect("home")
     return render(request, "blog_app/comments.html", {"form" : form})
+
+#----------------- Admin section --------------------------
+@login_required(login_url="log-in")
+@admin_only
+def view_dashboard(request):
+    user = Profile.objects.all().order_by('-created_at')
+    blog = Blog.objects.all().select_related('author', 'author__profile').order_by('-created_at')
+    comment = Comment.objects.all()
+    cont_dict = {
+        "users" : user,
+        "blogs" : blog,
+        "comment" : len(comment),
+        "total_user" : len(user),
+        "total_blog" : len(blog),
+    }
+    return render(request, "admin/dashboard.html", cont_dict)
+
+@login_required(login_url="log-in")
+@admin_only
+def change_status(request, id):
+    user = get_object_or_404(Profile, pk=id)
+    if user.is_active:
+        user.is_active = False
+        user.save()
+        messages.success(request, "Account blocked")
+    else:
+        user.is_active = True
+        user.save()
+        messages.success(request, "Account active")
+    return redirect("dashboard")
+
+@login_required(login_url="log-in")
+@admin_only
+def approve_post(request, id):
+    blog = get_object_or_404(Blog, pk=id)
+    if blog.approved:
+        blog.approved = False
+        blog.save()
+        messages.success(request, "Blog not approved")
+    else:
+        blog.approved = True
+        blog.save()
+        messages.success(request, "Blog approved")
+    return redirect("dashboard")
